+++ semcore/spin-container/src/SpinContainer.jsx
@@ -32,12 +32,14 @@ class SpinContainerRoot extends Component {
 
   render() {
     const SSpinContainer = Root;
-    const { styles, Children, loading } = this.asProps;
+    const { styles, Children, loading, forcedAdvancedMode } = this.asProps;
 
-    const advancedMode = isAdvanceMode(Children, [
-      SpinContainer.Content.displayName,
-      SpinContainer.Overlay.displayName,
-    ]);
+    const advancedMode =
+      forcedAdvancedMode ||
+      isAdvanceMode(Children, [
+        SpinContainer.Content.displayName,
+        SpinContainer.Overlay.displayName,
+      ]);
 
     return sstyled(styles)(
       <SSpinContainer render={Box} aria-busy={loading}>