+++ semcore/tooltip/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [6.5.0] - 2023-08-24
+
+### Added
+
+- Internal API for enforcing component advanced mode.
+
 ## [6.4.1] - 2023-08-24
 
 ### Changed